A pessimized implementation of thread-safe calls
Since the call closures _may_ be invoked in the context of another task,
it's a bad idea to pass them mutable pointers to on-stack boxes.

The wrapped objects now hold references to the default GMainContext
of the thread they were obtained in. The context pointer is woven
through the Rust <-> GLib interface to ensure that no objects are
used outside of their GMainContext. If needed, a call is dispatched
through the GMainContext in another thread.