Serialize the test run
It's pretty bad form that grust::init() is not reentrant.
But even though that should be fixed, running tests in parallel
creates unpredictability in on-stack vs. off-stack paths that
contest the default GMainContext. I should think about isolating
the tests better, as surely parallel test runs are a good stress
test for thread safety.