Various test suite improvements
Factored out the spawn_with_future function. It may be handy
for some tests to wait on a task result.
Renamed some tests to better correspond to their subject matter:
we are not necessarily hitting the off-stack path in all cross-task
tests.
Bring off_task_as_interface to what it was supposed to do.
Sundry stylistic touches.