Refactored MainLoop and related APIs
Added a RAII wrapper LoopRunner around creating a GMainContext,
pushing it as the thread default, and creating a GMainLoop to run it.
The MainLoop wrapper can be reffed out for closures and such.

While Rust RFC 0062 (removal of non-native runtimes) hasn't landed,
MainLoop and LoopRunner are kept in the module native as a
reminder that they are only safe to use with the native runtime.