Reverted to the run_after method for LoopRunner
The idea is to ensure that the API user can only access the
MainLoop and the MainContext after those have been properly set up,
and also they can be sure to invoke async closures in the thread-default
context of the loop as long as they do it in the closure.

With the thread context setup done implicitly when calling
LoopRunner::new(), it was possible to get confusing behavior by
inadvertently swapping creation of the runner and issuing async operations.

The users can still get their callbacks to never be called by
invoking async methods outside run_after, but it's not a program safety
issue, and it should be familiar to any Gio programmer.