Use markers in C structures to prohibit Copy and decide on Send/Sync
The structures returned by GLib-style APIs cannot be safely copied,
so a marker is put inside each structure which prohibits Copy.
marker::ObjectMarker is used for most objects which are not
Send or Sync either. We prohibit Send, not only Sync, because
some object state may be internally aliased between different
objects.
marker::SyncObjectMarker is used for threadsafe objects.

This allows us to remove the Threadsafe trait.