Don't require Send bound on closures passed to thread-unsafe methods
The rationale is, the callback is almost certainly invoked in the same
thread as where the object "is". Having a reference to a thread-unsafe
object in a thread concurrent to the GMainContext thread where it is used
is an unsafe situation to begin with. The closure dispatched in a
different thread is then just another consequence of the safety breach.
The exceptions should be rare; for g_source_set_callback() and similar
odd cases we'll either provide bespoke bindings in Grust or come up
with special annotations.

One possibility that can't be eliminated is when a thread in
non-Rust code acquires our thread's context and runs it.
Though concurrency is preempted, this can still screw up TLS
and possibly Rust runtime internals. But the same applies to
non-Rust code making assumptions about thread-local dispatch,
so the whole concept of stealing contexts should be an unsafe anti-pattern.

Allowing non-Send closures takes away the pain of not being able to
safely capture references to the majority of object types in callback
closures.