Disallow mutable references to objects
Object structs are views into larger privately allocated objects.
If mutable references to them can be obtained in safe code, the contents
can be overwritten with e.g. std::mem::swap without requiring an unsafe
block, resulting in undefined behavior. So we use only immutable
references in the bindings, regardless of constness of the
pointers in the bound functions.

For the hopefully rare cases where a public data member is available in
an object structure, a safe accessor can be generated providing
interior mutability. The bindings should never read data members
directly from the raw structures, preserving soundness for the
optimizer.